Refactored component css structure, renamed img folder to assets, updated logo
