+++ src/App.css
@@ -1,0 +1,5 @@
+.App {
+    background: rgb(236, 236, 236);
+    width: 100vw;
+    height: 100vh;
+}
