+++ src/components/Home/Home.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import './Home.scss';
+import News from '../News/News';
+import Weather from '../Weather/Weather';
+import Upcoming from '../Upcoming/Upcoming';
+
+function Home() {
+    return (
+        <div className="home-container">
+            <News />
+            <Weather/>
+            <Upcoming />
+        </div>
+    )
+}
+
+export default Home;
