+++ src/components/Home/Home.scss
@@ -1,0 +1,17 @@
+.home-container {
+    overflow-y: hidden;
+    overflow-x: scroll;
+    scrollbar-width: none;
+    -ms-overflow-style: none;
+    display: grid;
+    grid-auto-columns: minmax(440px, 40vw) minmax(700px, 60vw);
+    grid-auto-rows: 50vh 30vh 5vh;
+    grid-template-areas:
+        "news weather"
+        "news upcoming"
+        ". upcoming"
+}
+
+.home-container::-webkit-scrollbar {
+    display: none;
+}
