+++ src/components/TakeABreak/TakeABreak.scss
@@ -1,0 +1,6 @@
+.break-container {
+    margin-top: 2rem;
+    text-align: center;
+    font-weight: 600;
+    font-size: 36px;
+}
