+++ src/components/Topbar/Topbar.scss
@@ -1,0 +1,23 @@
+.topbar-container {
+    background-color: rgb(73, 73, 73);
+    height: 90px;
+    width: 100%;
+    font-size: 20px;
+    display: flex;
+
+    .topbar-logo {
+        padding: 7px 0;
+        height: 66px;
+    }
+
+    .topbar-item {
+        display: flex;
+        align-items: center;
+        height: 80px;
+    }
+
+    .topbar-button {
+        margin-left: 0.5rem;
+        margin-right: 0.5rem;
+    }
+}
