+++ src/components/Upcoming/Upcoming.css
@@ -1,0 +1,91 @@
+.upcoming-container {
+    grid-area: upcoming;
+    margin-left: 4vw;
+    position: relative;
+}
+
+.upcoming-header {
+    text-align: center;
+    font-family: 'Courier New', Courier, monospace;
+    font-size: 36px;
+    font-weight: 500;
+    padding-top: 10vh;
+}
+
+.upcoming-button {
+    margin-left: 2rem;
+}
+
+.upcoming-card-list {
+    display: flex;
+    position: absolute;
+    top: 25vh;
+    left: 5vw;
+}
+
+.upcoming-card {
+    height: max(325px, 40vh);
+    width: max(250px, 20vw);
+    padding: 0.4rem 0.5rem 0 1.2rem;
+    border-radius: 15px;
+    background: rgb(219,219,219);
+    box-shadow: -0.5rem 0.75rem 1.25rem #000;
+    transition: 0.5s;
+    font-family: 'Courier New', Courier, monospace;
+    overflow: scroll;
+    scrollbar-width: none;
+    -ms-overflow-style: none;
+    transform: scale(1);
+}
+
+.upcoming-card::-webkit-scrollbar {
+    display: none;
+}
+
+.upcoming-card:hover {
+    transform: translateY(min(-275px, -30vh));
+    background: rgb(236, 224, 224);
+}
+
+.upcoming-card:hover~.upcoming-card {
+    transform: translateX(max(125px, 10vw));
+}
+
+.upcoming-card:not(:first-child) {
+    margin-left: min(-125px, -10vw);
+}
+
+.due-soon {
+    background: rgb(236, 214, 214);
+    transform: translateY(min(-23px, -2.5vh));
+}
+
+.due-soon:hover {
+    background: rgb(236, 199, 199);
+}
+
+.upcoming-card-header {
+    height: max(50px, 8vh);
+    border-bottom: solid 2px;
+    margin-bottom: 0.5rem;
+    overflow: hidden;
+    font-size: 2.5vh;
+}
+
+.upcoming-card-header > * {
+    margin: 0;
+}
+
+.upcoming-card-body {
+    font-size: 2.5vh;
+}
+
+.upcoming-card-button {
+    margin-bottom: 1rem;
+}
+
+@media screen and (max-height: 500px) {
+    .upcoming-container {
+        display: none;
+    }
+}
