+++ src/components/Workmap/Workmap.scss
@@ -1,0 +1,21 @@
+.workmap-container {
+    width: 100vw;
+    height: 85vh;
+    text-align: center;
+    overflow: hidden;
+}
+
+.workmap-header {
+    margin: 1rem 0.5rem 1rem 0.5rem;
+    font-family:'Courier New', Courier, monospace;
+    font-size: 36px;
+}
+
+.workmap-header-icon {
+    margin-left: 1rem;
+    height: 36px;
+}
+
+.workmap-header-icon:hover {
+    transform: scale(1.2, 1.2);
+}
