--- src/services/quotes-service.js
@@ -1,8 +1,0 @@
-import axios from 'axios';
-
-// API from https://github.com/lukePeavey/quotable
-const apiUrl = "https://api.quotable.io";
-
-export const getRandomQuote = () => {
-    return axios.get(`${apiUrl}/random`);
-}
