--- src/styles/App.css
@@ -1,6 +1,0 @@
-.App {
-    background-image: url("../img/background.png");
-    background-repeat: repeat;
-    min-height: 100vh;
-    min-width: 100vw;
-}
