--- src/styles/Components.css
@@ -1,79 +1,0 @@
-.topbar-container {
-    height: 15vh;
-    font-size: 2.7vh;
-    display: flex;
-    justify-content: center;
-}
-
-.topbar-logo {
-    height: 8vh;
-}
-
-.topbar-button {
-    margin-left: 0.5rem;
-    margin-right: 0.5rem;
-}
-
-.landing-container {
-    position: relative;
-    text-align: center;
-}
-
-.landing-img {
-    margin-top: 15vh;
-    height: 25vh;
-}
-
-.landing-searchbar {
-    position: absolute;
-    left: 25%;
-    width: 50%;
-}
-
-.home-container {
-    overflow-y: hidden;
-    overflow-x: scroll;
-    scrollbar-width: none;
-    -ms-overflow-style: none;
-    height: 85vh;
-    display: grid;
-    grid-auto-columns: minmax(440px, 40vw) minmax(700px, 60vw);
-    grid-auto-rows: 50vh 30vh 5vh;
-    grid-template-areas:
-        "news weather"
-        "news upcoming"
-        ". upcoming"
-}
-
-.home-container::-webkit-scrollbar {
-    display: none;
-}
-
-.workmap-container {
-    width: 100vw;
-    height: 85vh;
-    text-align: center;
-    overflow: hidden;
-}
-
-.workmap-header {
-    margin: 1rem 0.5rem 1rem 0.5rem;
-    font-family:'Courier New', Courier, monospace;
-    font-size: 36px;
-}
-
-.workmap-header-icon {
-    margin-left: 1rem;
-    height: 36px;
-}
-
-.workmap-header-icon:hover {
-    transform: scale(1.2, 1.2);
-}
-
-.break-container {
-    margin-top: 30vh;
-    text-align: center;
-    font-weight: 600;
-    font-size: 36px;
-}
