--- src/styles/HomeComponents.css
@@ -1,241 +1,0 @@
-.news-container {
-    grid-area: news;
-    margin: 5vh 0 0 1.5vw;
-    box-shadow: -10px 12px 12px -2px #888888;
-    border: solid;
-    border-width: 3px;
-    border-radius: 8px;
-    border-color: rgb(233, 233, 233);
-    background-image: url('../img/newspaper.jpg');
-    background-repeat: repeat;
-    overflow: scroll;
-    scrollbar-width: none;
-    -ms-overflow-style: none;
-    transform: skewX(2deg);
-}
-
-.news-container::-webkit-scrollbar {
-    display: none;
-}
-
-.news-container:hover {
-    box-shadow: -8px 8px 10px -2px rgb(201, 201, 201);
-    transform: skewX(1deg);
-}
-
-.news-top {
-    text-align: center;
-    font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
-
-}
-
-.news-item {
-    display: block;
-}
-
-.news-img {
-    display: inline;
-    float: left;
-    opacity: 75%;
-}
-
-.news-title {
-    display: block;
-    font-weight: 400;
-}
-
-.news-description {
-    color: black;
-    float: left;
-}
-
-@keyframes shiftingGradient {
-    0% { background-position: 0% 50% }
-    50% { background-position: 100% 50% }
-    100% { background-position: 0% 50% }
-}
-
-.weather-container {
-    grid-area: weather;
-    height: 275px;
-    width: 700px;
-    margin: auto;
-    border: solid 2px;
-    border-radius: 5px;
-    display: grid;
-    grid-template-columns: 30% 8% 27% 8% 27%;
-    grid-template-rows: 100%;
-    grid-template-areas: "currWeather hourlyHeader hourlyWeather dailyHeader dailyWeather";
-    overflow: hidden;
-    white-space: nowrap;
-}
-
-.current-weather-container {
-    grid-area: currWeather;
-    text-align: center;
-    font-weight: 500;
-    color: rgb(0, 0, 0);
-}
-
-.hourly-weather-container {
-    grid-area: hourlyWeather;
-    padding-left: 0.5rem;
-    font-weight: 400;
-    color: rgb(0, 0, 0);
-}
-
-.daily-weather-container {
-    grid-area: dailyWeather;
-    padding-left: 0.5rem;
-    font-weight: 400;
-    color: rgb(0, 0, 0);
-}
-
-.weather-shifting-bg {
-    animation: shiftingGradient 120s ease infinite;
-    background-size: 200% 200%;
-}
-
-.weather-clear-bg {
-    background: url('../img/clear.jpg');
-}
-
-.weather-cloud-bg {
-    background: url('../img/cloud.jpg');
-}
-
-.weather-rain-bg {
-    background: url('../img/rain.jpg');
-}
-
-.weather-snow-bg {
-    background: url('../img/snow.jpg');
-}
-
-.weather-mist-bg {
-    background: url('../img/mist.jpg');
-}
-
-.hourly-header {
-    grid-area: hourlyHeader;
-    background: #369;
-    height: 100%;
-}
-
-.hourly-header-text {
-    font: bold 20px Sans-Serif;
-    letter-spacing: 2px;
-    text-transform: uppercase;
-    color: #fff;
-    position: relative;
-    top: 53%;
-    transform: rotate(270deg);
-}
-
-.daily-header {
-    grid-area: dailyHeader;
-    background: #369;
-    height: 100%;
-}
-
-.daily-header-text {
-    font: bold 20px Sans-Serif;
-    letter-spacing: 2px;
-    text-transform: uppercase;
-    color: #fff;
-    position: relative;
-    top: 48%;
-    transform: rotate(270deg);
-}
-
-.upcoming-container {
-    grid-area: upcoming;
-    margin-left: 4vw;
-    position: relative;
-}
-
-.upcoming-header {
-    text-align: center;
-    font-family: 'Courier New', Courier, monospace;
-    font-size: 36px;
-    font-weight: 500;
-    padding-top: 10vh;
-}
-
-.upcoming-button {
-    margin-left: 2rem;
-}
-
-.upcoming-card-list {
-    display: flex;
-    position: absolute;
-    top: 25vh;
-    left: 5vw;
-}
-
-.upcoming-card {
-    height: max(325px, 40vh);
-    width: max(250px, 20vw);
-    padding: 0.4rem 0.5rem 0 1.2rem;
-    border-radius: 15px;
-    background: rgb(219,219,219);
-    box-shadow: -0.5rem 0.75rem 1.25rem #000;
-    transition: 0.5s;
-    font-family: 'Courier New', Courier, monospace;
-    overflow: scroll;
-    scrollbar-width: none;
-    -ms-overflow-style: none;
-    transform: scale(1);
-}
-
-.upcoming-card::-webkit-scrollbar {
-    display: none;
-}
-
-.upcoming-card:hover {
-    transform: translateY(min(-275px, -30vh));
-    background: rgb(236, 224, 224);
-}
-
-.upcoming-card:hover~.upcoming-card {
-    transform: translateX(max(125px, 10vw));
-}
-
-.upcoming-card:not(:first-child) {
-    margin-left: min(-125px, -10vw);
-}
-
-.due-soon {
-    background: rgb(236, 214, 214);
-    transform: translateY(min(-23px, -2.5vh));
-}
-
-.due-soon:hover {
-    background: rgb(236, 199, 199);
-}
-
-.upcoming-card-header {
-    height: max(50px, 8vh);
-    border-bottom: solid 2px;
-    margin-bottom: 0.5rem;
-    overflow: hidden;
-    font-size: 2.5vh;
-}
-
-.upcoming-card-header > * {
-    margin: 0;
-}
-
-.upcoming-card-body {
-    font-size: 2.5vh;
-}
-
-.upcoming-card-button {
-    margin-bottom: 1rem;
-}
-
-@media screen and (max-height: 500px) {
-    .upcoming-container {
-        display: none;
-    }
-}
